requirements.txt
python-dotenv
asyncpg~=0.30.0
fastapi~=0.121.1
